+++ src/nice_int/nice_u32.rs
@@ -145,6 +145,38 @@ impl NiceU32 {
 		out.parse(num);
 		out
 	}
+
+	/// # Replace.
+	///
+	/// Reuse the backing storage behind `self` to hold a new nice number.
+	///
+	/// ## Examples.
+	///
+	/// ```
+	/// use dactyl::NiceU32;
+	///
+	/// let mut num = NiceU32::from(3141592653_u32);
+	/// assert_eq!(num.as_str(), "3,141,592,653");
+	///
+	/// num.replace(12345);
+	/// assert_eq!(num.as_str(), "12,345");
+	/// ```
+	///
+	/// Note that custom separators, if any, are preserved.
+	///
+	/// ```
+	/// use dactyl::NiceU32;
+	///
+	/// let mut num = NiceU32::with_separator(3141592653_u32, b'_');
+	/// assert_eq!(num.as_str(), "3_141_592_653");
+	///
+	/// num.replace(12345);
+	/// assert_eq!(num.as_str(), "12_345");
+	/// ```
+	pub fn replace(&mut self, num: u32) {
+		self.from = SIZE;
+		self.parse(num);
+	}
 }
 
 
@@ -189,6 +221,7 @@ mod tests {
 		assert_eq!(two.cmp(&one), std::cmp::Ordering::Greater);
 
 		// Check a subset of everything else.
+		let mut last = NiceU32::empty();
 		let mut rng = fastrand::Rng::new();
 		for i in std::iter::repeat_with(|| rng.u32(..)).take(SAMPLE_SIZE) {
 			let nice = NiceU32::from(i);
@@ -199,7 +232,16 @@ mod tests {
 			assert_eq!(nice.len(), nice.as_str().len());
 			assert_eq!(nice.len(), nice.as_bytes().len());
 			assert!(! nice.is_empty());
+
+			// Replacement should yield the same thing.
+			assert_ne!(nice, last);
+			last.replace(i);
+			assert_eq!(nice, last);
 		}
+
+		// Make sure back to zero works.
+		last.replace(0);
+		assert_eq!(last.as_str(), "0");
 	}